requirements.txt
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.0
psycopg2==2.9.9
sqlparse==0.4.4
